MIPS: Remove the section table when CONFIG_EMBEDDED=y
The section table is useless for the run-time executable, so, can be
removed completely.

A c version of sstrip is added from
https://dev.openwrt.org/browser/trunk/tools/sstrip/src/sstrip.c instead.

Because the c version removes more 0 bits before the section table, it
saves more bits, so, use it instead of mine, but mine is simpler if time
allows, I will let it do more things, such like removing the necessary
bits.

At last, add a kernel config optoin to allow people disable it if they
need the section table for debugging.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>