param: Allow setup a list of kernel parameters
when CONFIG_SETUP_PARAM is not set, this allow users to config a list of
kernel parameters to be setup, this may be very helpful for the embedded
system, which only need a fixed list kernel parameters.

BTW, __raw_early_param is added for the raw version of early_param to
cope with the early param setup.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>