MIPS: Add another sstrip from openwrt and make it configurable
With my old sstrip.sh, it doesn't remove the section table information
in the ELF header and make the loader confused. This fixes it.

And at the same time, a new c version of sstrip is added from
https://dev.openwrt.org/browser/trunk/tools/sstrip/src/sstrip.c instead.

Because the c version removes more 0 bits before the section table, it
saves more bits, so, use it instead of mine, but mine is simpler if time
allows, I will let it do more things, such like removing the necessary
bits.

At last, add a kernel config optoin to allow people disable it if they
need the section table for debugging.

NOTE: Do we need to remove the section table for the vmlinux.bin, does
objcopy do it already? If not, we may need to do it.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>