Module: Optimize linker scripts (FIXME)
1. Reduce number of sections via merging similar sections

Note: This is because the regular expression doesn't work, we use *
instead but this may include the a..b sections.... This need to be fixed
in the future.

2. Discard more sections(for example, the sections will never be used or
will be used for debugging)

Note: only MIPS is tested and we don't consider debugging this time,
we'd better KEEP the sections for debugging.

In order to add your arch specific support, please put a module.lds in
your ARCH's directory: arch/ARCH/module.lds, please use
arch/mips/module.lds as a reference.

3. TODO: Add gc sections support for modules, need to KEEP the root
sections and discard the others by gc-sections automatically.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>