X86: Make sys_ptrace configurable
The xstate* and regs_query_register* related parts should be taken care
of next time, they are 'fixed up' temporarily for compiling, the current
fixup is very very ugly.

BTW, the calling to do_setitimer() is fixed up through defining it to
empty. Need to take care of return a better value for an undefined
function.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>