gc-sections/powerpc: Tmp Fixup of the pattern support of the old ld
[ Fix MIPS, ARM and X86, we already have the new ld 2.21 for testing,
so, don't add this workaround currently, but in the future, we may need
to check the version of ld and then determine to add this workaround,
   otherwise, the kernel image generated by gc-sections will be bigger
   for the .text.* sections may be linked to their own section in the
   executables and the alignment may take lots of space ]

The old ld (2.20.1.20100303) doesn't support the pattern like:

*(.text.[A-Za-z0-9_$^]*)

	or

*(.text.[^.]*)

Therefore, *(.sbss.*) must be used, but this pattern may merge some
magic sections defined by the kernel itself, such as *(.text..XX.YY)

In order to avoid merging the magic sections (e.g .data..*) to .data, we
move them before the section of .data and then they will be linked to
their own sections and the .data.* pattern will not relink them. this
may be a good solution to the bug of the ld.

Note: 2.21.0.20110327 does support the above pattern for ARM, but still
need to check it for powerpc.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>