gc-sections: add ld-version.sh to detect the version of ld
The old ld version (<2.21 or may 2.20.50?) doesn't support the pattern
like:

*(.text.[A-Za-z0-9_$^]*) or *(.text.[^.]*)

Which prevent the linker to merge the sections generated by
-ffunction-sections, -fdata-sections and our new defined __section*,
__asm_section* to the last .text section and will generate very large
executable, therefore, the feature of our -ffunction-sections ... should
only be enabled when the ld version used is >= 2.21.

The coming version will add the detection of the ld version and then
make decision.

Note: --gc-sections doesn't depend on -ffunction-sections,
-fdata-sections, .... therefore, it is able to be enabled all the time.

Signed-off-by: Wu Zhangjin <wuzhangjin@gmail.com>